backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "clerk-backend-api>=3.3.0",
    "crewai[tools]>=0.175.0",
    "fastapi>=0.116.1",
    "google-generativeai>=0.8.5",
    "openai>=1.102.0",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.35.0",
]
